pyproject.toml
[project]
name = "bmad-dev-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.71",
    "python-dotenv>=1.2.2",
]
